requirement.txt
alabaster==0.7.13
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.2.2
asgiref==3.7.2
attrs==23.2.0
Babel==2.13.0
bcrypt==4.1.2
beautifulsoup4==4.10.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
chroma-py==0.1.0.dev1
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
colorlog==6.8.2
colour==0.1.5
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-versatileimagefield==3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
docutils==0.20.1
drf-flex-fields==1.0.2
exceptiongroup==1.1.3
fabric==3.2.2
filelock==3.13.1
flake8==6.1.0
flasgger==0.9.7.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.49.0
geojson==3.1.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
googlemaps==4.10.0
greenlet==3.0.1
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
idna==3.4
imagesize==1.4.1
importlib-metadata==4.6.4
iniconfig==2.0.0
invoke==2.2.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.0.1
jsonschema-specifications==2023.12.1
kaleido==0.2.1
keyring==23.5.0
Kivy==2.3.0
Kivy-Garden==0.1.5
kivy-garden.graph==0.4.0
kivy-garden.mapview==1.0.6
kivy_matplotlib_widget==0.10.0
KivyCalendar==0.1.3
KivyGradient==0.0.4
kivymd==1.2.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
mapboxgl==0.10.2
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib==3.8.3
mccabe==0.7.0
mistune==3.0.2
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
nox==2023.4.22
numpy==1.26.4
oauthlib==3.2.0
openai==1.12.0
packaging==23.2
paramiko==3.4.0
pathlib2==2.3.7.post1
pep8==1.7.1
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.3.0
plyer==2.1.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.5.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pyflakes==3.1.0
Pygments==2.16.1
PyGObject==3.42.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyJWT==2.3.0
pymongo==4.6.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyrsistent==0.20.0
pytest==7.4.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
PyYAML==5.4.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==1.4.22
sqlparse==0.4.4
systemd-python==234
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.2
ttystatus==0.38
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.25.0
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==1.0.0
